feat: Cache remote objects and write to config
So I'm not totally happy with this design, I worked on this a bunch yesterday, then slept on it. Not really sure what I want here. Having a local cache of our remote state is what we want, but I'm not sure it's helpful beyond this developer's experience.

Basically we're holding our remote objects, and doing a dance with the config generated objects. I was doing a bunch of basemodel to dict conversions, but decided just to copy the objects. Our endpoint returns a list, which we'll need to paginate eventually, but we need to maintain a mapping of objects. Maybe not, all we want is to be able to lookup objects by name or id. I'm not sure what the point is here.

Anyways, we should be able to instantiate the assistant handler within the app, and then lookup our assistant by name. No more hard coded IDs. From there we can use the default file store similarly to the new file handler. Then we should be ready to pilot.