tests: fix assistant_handler.py
We're running into issues with the OAI client handling again. We had a mismatch between the test env and the handler, which was init it's own with who knows what env file. Now, the Assistant handler takes an optional client parameter.

We're also phasing out the use of OAI IDs for this, although we do need one for the tests still.